requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
facebook-sdk==2.0.0
idna==2.6
matplotlib==2.0.2
numpy==1.13.1
prettytable==0.7.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
